+++ src/main/java/com/woorifisa/backend/subscription/dto/ProductPayDTO.java
@@ -1,0 +1,19 @@
+package com.woorifisa.backend.subscription.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+@Getter
+@Setter
+@Builder
+public class ProductPayDTO {
+    String prodName;
+    int prodPrice;
+}
